_site/requirements/python.txt
BeautifulSoup4
bibtexparser
pyyaml
